Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/hr-time/monotonic-clock.any-expected.txt


PASS self.performance.now() returns a positive number
PASS self.performance.now() difference is not negative
